pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tsbm"
version = "0.1.0"
description = "Production-grade time-series database benchmarking suite for QuestDB, CrateDB, and TimescaleDB"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.9",
    "rich>=13.0",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.3",
    "tomli>=2.0; python_version < '3.11'",
    "pyarrow>=12.0",
    "pandas>=2.0",
    "numpy>=1.23",
    "hdrhistogram>=0.10.2",
    "psutil>=5.9",
    "asyncpg>=0.28",
    "aiohttp>=3.8",
    "python-dateutil>=2.8",
    "orjson>=3.8",
    "tzdata>=2023.3; sys_platform == 'win32'",
]

[project.optional-dependencies]
questdb     = ["questdb>=1.0"]
cratedb     = ["crate>=1.0"]
timescaledb = []
dashboard   = ["streamlit>=1.28", "plotly>=5.0"]
all         = ["tsbm[questdb,cratedb,timescaledb]"]
dev         = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "mypy>=1.0",
]

[project.scripts]
tsbm = "tsbm.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/tsbm"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]
markers      = ["integration: requires running database containers"]

[tool.mypy]
python_version = "3.10"
strict         = true
